Add Azure deployment config and embed Angular in Spring Boot
- Remove server.servlet.context-path, add WebConfig /api prefix for RestControllers
- Embed Angular dist in src/main/resources/static (built by CI)
- Add SpaController for Angular client-side routing fallback
- Update SecurityConfig for new URL patterns
- Add application-prod.properties with Azure port config
- Add GitHub Actions workflow for CI/CD to Azure App Service

+++ src/main/java/com/poly/dindor/repository/OrigineRepository.java
@@ -4,5 +4,5 @@ import com.poly.dindor.entity.Origine;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrigineRepository extends JpaRepository<Origine, Long> {
-    boolean existsByCode(Integer code);
+    boolean existsByCode(Long code);
 }
